Rearranged folders structure. Fixed bugs
