+++ aeronet_raster/geoobject/geoobject.py
