+++ docs/conf.py
@@ -143,7 +143,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, 'aeronet', u'Aeronet Documentation',
+    (master_doc, 'aeronet_raster', u'Aeronet Documentation',
      [author], 1)
 ]
